+++ game/pieces.py
@@ -140,7 +140,31 @@ class Pawn(Pieces):
             print(e)
             raise
 
+class Knight(Pieces):
+    def limit(self, new_position):
+        row = self.__position__[0]
+        column = self.__position__[1]
+        list = [[row - 2, column - 1],
+                [row - 2, column + 1],
+                [row - 1, column - 2],
+                [row - 1, column + 2],
+                [row + 1, column - 2],
+                [row + 1, column + 2],
+                [row + 2, column - 1],
+                [row + 2, column + 1]]
         
+        if new_position not in list:
+            raise LimitedMove()
+    
+    def movement(self, new_position):
+        try:
+            self.on_board(new_position)
+            self.limit(new_position)
+        except Exception as e:
+            print(e)
+            raise
+    
+